+++ game-warhammer40kdarkheresy/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None Planned
 
+## [0.1.1] - 2026-07-25
+
+- Fixed: Using "Update all" to update several mods at once could move them in the load order or uncheck them.
+
 ## [0.1.0] - 2026-02-16
 
 - Inital Release